requirements.txt
asgiref==3.8.1
Django==4.2.21
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
googletrans==4.0.0-rc1
psycopg2-binary==2.9.10
pillow==11.2.1
python-dotenv==1.1.0
pyzbar==0.1.9
Markdown==3.8
bleach==6.2.0
